--- src/main/java/LBot/exception/ExecuteCommandException.java
@@ -1,12 +1,0 @@
-package LBot.exception;
-
-/**
- * This exception is thrown when the chatbot encounters issues executing a command.
- * <p>
- * For example, if an incorrect command keyword is provided.
- */
-public class ExecuteCommandException extends LBotException {
-    public ExecuteCommandException(String message) {
-        super(message);
-    }
-}
