--- src/main/java/LBot/exception/LBotException.java
@@ -1,12 +1,0 @@
-package LBot.exception;
-
-public abstract class LBotException extends Exception {
-    public LBotException(String message) {
-        super(message);
-    }
-
-    @Override
-    public String toString() {
-        return getMessage();
-    }
-}
