+++ src/main/java/lbot/exception/LBotException.java
@@ -1,0 +1,15 @@
+package lbot.exception;
+
+/**
+ * This is the abstract parent LBotException class.
+ */
+public abstract class LBotException extends Exception {
+    public LBotException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String toString() {
+        return getMessage();
+    }
+}
