+++ src/main/java/lbot/exception/ParseCommandException.java
@@ -1,0 +1,13 @@
+package lbot.exception;
+
+/**
+ * This exception is thrown when the chatbot
+ * encounters issues parsing command arguments.
+ * <p>
+ * For example, the provided datetime format is incorrect.
+ */
+public class ParseCommandException extends LBotException {
+    public ParseCommandException(String message) {
+        super(message);
+    }
+}
